Fixed a bug where emitted events would fire multiple times

+++ frontend/src/windows/main/ts/roblox/instance.ts
@@ -1,12 +1,12 @@
-import { events, filesystem, os } from "@neutralinojs/lib";
-import { isProcessAlive, removeNonUTF8CharactersFromString } from "../utils";
-import { sleep } from "../utils";
-import path from "path-browserify";
-import { getRobloxPath } from "./path";
-import Roblox from ".";
+import { events, filesystem, os } from '@neutralinojs/lib';
+import { isProcessAlive, removeNonUTF8CharactersFromString } from '../utils';
+import { sleep } from '../utils';
+import path from 'path-browserify';
+import { getRobloxPath } from './path';
+import Roblox from '.';
 
 type EventHandler = (data?: any) => void;
-type Event = "exit" | "gameInfo" | "gameEvent";
+type Event = 'exit' | 'gameInfo' | 'gameEvent';
 export interface GameEventInfo {
 	event: string;
 	data: string;
@@ -20,44 +20,44 @@ interface Entry {
 // code adapted from https://github.com/pizzaboxer/bloxstrap/blob/main/Bloxstrap/Integrations/ActivityWatcher.cs
 const Entries: Entry[] = [
 	{
-		event: "GameJoining",
-		match: "[FLog::Output] ! Joining game",
+		event: 'GameJoining',
+		match: '[FLog::Output] ! Joining game',
 	},
 	{
-		event: "GameStartJoining",
-		match: "[FLog::SingleSurfaceApp] launchUGCGameInternal"
+		event: 'GameStartJoining',
+		match: '[FLog::SingleSurfaceApp] launchUGCGameInternal',
 	},
 	{
-		event: "GameJoiningPrivateServer",
-		match: "[FLog::GameJoinUtil] GameJoinUtil::joinGamePostPrivateServer",
+		event: 'GameJoiningPrivateServer',
+		match: '[FLog::GameJoinUtil] GameJoinUtil::joinGamePostPrivateServer',
 	},
 	{
-		event: "GameJoiningReservedServer",
-		match: "[FLog::GameJoinUtil] GameJoinUtil::initiateTeleportToReservedServer",
+		event: 'GameJoiningReservedServer',
+		match: '[FLog::GameJoinUtil] GameJoinUtil::initiateTeleportToReservedServer',
 	},
 	{
-		event: "GameJoiningUDMUX",
-		match: "[FLog::Network] UDMUX Address = ",
+		event: 'GameJoiningUDMUX',
+		match: '[FLog::Network] UDMUX Address = ',
 	},
 	{
-		event: "GameJoined",
-		match: "[FLog::Network] serverId:",
+		event: 'GameJoined',
+		match: '[FLog::Network] serverId:',
 	},
 	{
-		event: "GameDisconnected",
-		match: "[FLog::Network] Time to disconnect replication data:",
+		event: 'GameDisconnected',
+		match: '[FLog::Network] Time to disconnect replication data:',
 	},
 	{
-		event: "GameTeleporting",
-		match: "[FLog::SingleSurfaceApp] initiateTeleport",
+		event: 'GameTeleporting',
+		match: '[FLog::SingleSurfaceApp] initiateTeleport',
 	},
 	{
-		event: "GameMessage",
-		match: "[FLog::Output] [BloxstrapRPC]",
+		event: 'GameMessage',
+		match: '[FLog::Output] [BloxstrapRPC]',
 	},
 	{
-		event: "GameLeaving",
-		match: "[FLog::SingleSurfaceApp] leaveUGCGameInternal",
+		event: 'GameLeaving',
+		match: '[FLog::SingleSurfaceApp] leaveUGCGameInternal',
 	},
 ];
 
@@ -68,25 +68,25 @@ interface Pattern {
 
 const Patterns: Pattern[] = [
 	{
-		event: "GameJoiningEntry",
+		event: 'GameJoiningEntry',
 		regex: /! Joining game '([0-9a-f\-]{36})' place ([0-9]+) at ([0-9\.]+)/g,
 	},
 	{
-		event: "GameJoiningUDMUX",
+		event: 'GameJoiningUDMUX',
 		regex: /UDMUX Address = ([0-9\.]+), Port = [0-9]+ \| RCC Server Address = ([0-9\.]+), Port = [0-9]+/g,
 	},
 	{
-		event: "GameJoinedEntry",
+		event: 'GameJoinedEntry',
 		regex: /serverId: ([0-9\.]+)\|[0-9]+/g,
 	},
 	{
-		event: "GameMessageEntry",
+		event: 'GameMessageEntry',
 		regex: /\[BloxstrapRPC\] (.*)/g,
 	},
 	{
-		event: "GameCrashEntry",
+		event: 'GameCrashEntry',
 		regex: /\[FLog::CrashReportLog\] (.*)/g,
-	}
+	},
 ];
 
 export class RobloxInstance {
@@ -95,9 +95,9 @@ export class RobloxInstance {
 	private gameInstance: number | null = null;
 	private latestLogPath: string | null = null;
 	private logsInstance: os.SpawnedProcess | null = null;
-	private lastLogs: string = "";
+	private lastLogs: string = '';
 	private isWatching: boolean = false;
-	private onEvent: Promise<events.Response> | null = null
+	private onEvent: Promise<events.Response> | null = null;
 
 	/** Adds a handler to an event */
 	public on(event: Event, handler: EventHandler) {
@@ -124,7 +124,7 @@ export class RobloxInstance {
 	}
 
 	watchLogs: boolean;
-	constructor(watch: boolean) {
+	constructor(watch: boolean, url?: string) {
 		this.watchLogs = watch;
 	}
 
@@ -134,22 +134,30 @@ export class RobloxInstance {
 	}
 
 	/** Starts the Roblox Instance */
-	public async start() {
-		if (this.gameInstance) throw new Error("An instance is already running");
+	public async start(url?: string) {
+		if (this.gameInstance) throw new Error('An instance is already running');
 
-		console.log("Opening Roblox instance");
+		console.log('Opening Roblox instance');
+
 		// Launch Roblox
-		await os.execCommand(`open "${getRobloxPath()}"`);
+		if (url) {
+			console.log(`Opening from URL: ${url}`);
+			await Roblox.Utils.toggleURI(false, false);
+			await os.execCommand(`open ${url}`);
+		} else {
+			await os.execCommand(`open "${getRobloxPath()}"`);
+		}
 
 		await sleep(1000);
+		await Roblox.Utils.toggleURI(true, false);
 
 		// We find every roblox processes and get the RobloxPlayer one
-		const robloxProcess = await (await os.execCommand('pgrep -f "Roblox"')).stdOut.split("\n");
+		const robloxProcess = await (await os.execCommand('pgrep -f "Roblox"')).stdOut.split('\n');
 		for (const pid of robloxProcess) {
 			const info = (await os.execCommand(`ps -p ${pid} -o command=`)).stdOut.trim();
 			if (info.length < 2) continue;
 			const processFileName = path.basename(info);
-			if (processFileName === "RobloxPlayer") {
+			if (processFileName === 'RobloxPlayer') {
 				this.gameInstance = parseInt(pid);
 			}
 		}
@@ -159,7 +167,7 @@ export class RobloxInstance {
 		}
 
 		// Find the latest log file
-		const logsDirectory = path.join(await os.getEnv("HOME"), "Library/Logs/Roblox");
+		const logsDirectory = path.join(await os.getEnv('HOME'), 'Library/Logs/Roblox');
 		let tries = 10;
 		while (this.latestLogPath == null) {
 			if (tries < 1) {
@@ -188,26 +196,49 @@ export class RobloxInstance {
 `);
 		console.log(`Logs watcher started with PID: ${this.logsInstance.pid}`);
 
-		this.onEvent = events.on("spawnedProcess", async (evt: CustomEvent) => {
-			// Check if the event comes from the logs watcher, and that it is stdOut
-			if (!this.isWatching || !this.logsInstance || evt.detail.id !== this.logsInstance.id) return;
-			if (evt.detail.action === "exit") {
-				console.log("Logs watcher exited with output: ")
-				console.log(evt.detail.data);
-				console.log("Restarting logs watcher");
-				await os.execCommand(`pkill -f "tail -f /Users/$(whoami)/Library/Logs/Roblox/"`);
-				await sleep()
-				this.logsInstance = await os.spawnProcess(`tail -f "${this.latestLogPath}" | while read line; do echo "Change"; done`);
-				return;
-			}
-			await os.execCommand(`iconv -f utf-8 -t utf-8 -c "${this.latestLogPath}" > /tmp/roblox_ablox.log`);
-			const content = (await os.execCommand(`cat /tmp/roblox_ablox.log`)).stdOut;
-			const newContent = content.split("\n").filter((l) => !this.lastLogs.includes(l));
-			await this.processLines(newContent);
-			this.lastLogs = content;
-		});
+		let isProcessing = false;
+		const handler = async (evt: CustomEvent) => {
+			// Ensure only one instance of the handler runs at a time
+			if (isProcessing) return;
+			isProcessing = true;
 
-		await this.processLines(content.split("\n"));
+			try {
+				// Check if the event comes from the logs watcher, and that it is stdOut
+				if (!this.isWatching || !this.logsInstance || evt.detail.id !== this.logsInstance.id) return;
+
+				if (evt.detail.action === 'exit') {
+					console.log('Logs watcher exited with output:', evt.detail.data);
+					console.log('Restarting logs watcher');
+
+					await os.execCommand(`pkill -f "tail -f /Users/$(whoami)/Library/Logs/Roblox/"`);
+					this.logsInstance = await os.spawnProcess(`tail -f "${this.latestLogPath}" | while read line; do echo "Change"; done`);
+					return;
+				}
+
+				// Convert the file to ensure proper encoding
+				await os.execCommand(`iconv -f utf-8 -t utf-8 -c "${this.latestLogPath}" > /tmp/roblox_ablox.log`);
+
+				// Read the content of the converted file
+				const content = (await os.execCommand(`cat /tmp/roblox_ablox.log`)).stdOut;
+
+				// Process only new lines
+				const contentLines = content.split('\n');
+				const newContent = contentLines.filter((line) => !this.lastLogs.includes(line));
+
+				if (newContent.length > 0) {
+					await this.processLines(newContent);
+					this.lastLogs = content;
+				}
+			} catch (error) {
+				console.error('Error processing log file:', error);
+			} finally {
+				isProcessing = false;
+			}
+		};
+		await events.off('spawnedProcess', handler);
+		events.on('spawnedProcess', handler);
+
+		await this.processLines(content.split('\n'));
 		this.lastLogs = content;
 		this.isWatching = true;
 
@@ -215,21 +246,20 @@ export class RobloxInstance {
 			// Check if instance is still alive
 			if (this.gameInstance && !(await isProcessAlive(this.gameInstance))) {
 				this.gameInstance = null;
-				this.emit("exit");
+				events.off('spawnedProcess', handler);
+				this.emit('exit');
 				await this.cleanup();
-				console.log("Instance is null, stopping.");
+				console.log('Instance is null, stopping.');
 				clearInterval(intervalId);
 			}
 		}, 500);
 	}
 
 	private processLines(lines: string[]) {
-		for (const entry of Entries)
-
 		for (const entry of Entries) {
 			const includedLines = lines.filter((line) => line.includes(entry.match));
 			for (const line of includedLines) {
-				this.emit("gameEvent", { event: entry.event, data: line });
+				this.emit('gameEvent', { event: entry.event, data: line });
 			}
 		}
 
@@ -238,7 +268,7 @@ export class RobloxInstance {
 			for (const line of matchedLines) {
 				const match = line.match(pattern.regex);
 				if (match) {
-					this.emit("gameEvent", { event: pattern.event, data: match[0] });
+					this.emit('gameEvent', { event: pattern.event, data: match[0] });
 				}
 			}
 		}
@@ -255,7 +285,7 @@ export class RobloxInstance {
 	public async quit() {
 		if (this.gameInstance == null) throw new Error("The instance hasn't be started yet");
 		await this.cleanup();
-		console.log("Quitting Roblox");
+		console.log('Quitting Roblox');
 		await os.execCommand(`kill -9 ${this.gameInstance}`);
 	}
 }
